idt/gdt: interface improvements
        This commit makes some minor changes to the interface that I
        hastily put together to use incomplete _info types. Now error
        codes are appropriately returned, and some types have changed to
        prevent overflow situations.

        There are still issues. For example, idt_entry_set should
        probably not uncritically accept input the way it currently
        does. idt_entry_add should probably give the caller an IDT entry
        descriptor which can be used for any purposes in the future. It
        also needs to search the bitmap for free entries once a
        deletion function has been added.

        In other words, the interface is not mature at the moment, but
        this at least gives me error detection.

+++ libk/idt.h
@@ -36,9 +36,8 @@ struct idt_info *idt_init(void);
 size_t idt_entries_nr(struct idt_info *);
 size_t idt_entries_max(struct idt_info *);
 
-size_t
-idt_entry_set(struct idt_info *, void (*)(void), uint16_t, uint8_t, uint8_t);
-size_t idt_entry_add(struct idt_info *, void (*)(void), uint16_t, uint8_t);
+int idt_entry_set(struct idt_info *, void (*)(void), size_t, uint16_t, uint8_t);
+int idt_entry_add(struct idt_info *, void (*)(void), uint16_t, uint8_t);
 
 void exception_idt_init(struct idt_info *entries);
 void irq_idt_init(struct idt_info *entries);
